backenddjango/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.3
drf-yasg==1.21.7
drf-spectacular==0.26.5
django-model-utils==4.3.1
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
channels==4.0.0
channels-redis==4.1.0
Pillow==10.1.0
python-dateutil==2.8.2
reportlab==4.0.7
WeasyPrint==60.1
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.28.64
pdfrw==0.4
